requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
asyncpg==0.29.0
attrs==23.2.0
cffi==1.16.0
faust-cchardet==2.1.19
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
multidict==6.0.5
pycares==4.4.0
pycparser==2.22
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
yarl==1.9.4
